Move get-pub-trade-chunk-payloads fn into separate helper
