+++ workers/loc.api/sync/transaction.tax.report/helpers/index.js
@@ -2,8 +2,10 @@
 
 const lookUpTrades = require('./look-up-trades')
 const getTrxMapByCcy = require('./get-trx-map-by-ccy')
+const getPubTradeChunkPayloads = require('./get-pub-trade-chunk-payloads')
 
 module.exports = {
   lookUpTrades,
-  getTrxMapByCcy
+  getTrxMapByCcy,
+  getPubTradeChunkPayloads
 }
